Remove modal, use dialog instead
