+++ src/fullPageBookingForm/Components/Common/CustomForm/CustomForm.stories.tsx
@@ -1,67 +1,15 @@
 /** @jsx h */
 import { h } from "preact";
-import { FormFieldType } from "../../../../typings/Event";
-import { CustomFieldType } from "../../../Typings/CustomForm";
 import {
-  CustomForm,
-  CustomFormProps,
-  PerAttendeeTypeProps,
-  PerOrderTypeProps,
-} from "./CustomForm";
+  defaultPerOrderFormProps,
+  defaultPerAttendeeFormProps,
+} from "../../../__mocks__/CustomForm";
+import { CustomForm, CustomFormProps } from "./CustomForm";
 
 export default {
   title: "Full Page Booking Form/Common/Custom Form",
   component: CustomForm,
   argTypes: {},
-};
-
-export const defaultFields: CustomFieldType[] = [
-  {
-    label: "First",
-    isRequired: true,
-    type: FormFieldType.Text,
-  },
-  {
-    label: "Last",
-    isRequired: true,
-    type: FormFieldType.Text,
-  },
-  {
-    label: "Snack Preference",
-    isRequired: false,
-    type: FormFieldType.Text,
-  },
-  {
-    label: "Shirt Size?",
-    isRequired: true,
-    type: FormFieldType.Select,
-    options: ["XS", "S"],
-  },
-];
-
-export const defaultPerOrderFormProps: PerOrderTypeProps = {
-  formValues: [
-    {
-      fields: defaultFields,
-      name: "Default Form",
-    },
-  ],
-  formTitle: "Title of Form",
-  formDescription: "Description of form",
-};
-
-export const defaultPerAttendeeFormProps: PerAttendeeTypeProps = {
-  formValues: [
-    {
-      fields: defaultFields,
-      name: "Per Attendee Form",
-    },
-  ],
-  removeVariantModal: {
-    isOpen: false,
-    removeVariant: () => {},
-    setIsRemoveVariantModalOpen: () => {},
-  },
 };
 
 const defaultArgs: CustomFormProps = {
